fixed deprecation message because of changes from Eclipse-Bug 50164
